backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.135.3
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
httptools==0.7.1
idna==3.11
limits==5.8.0
loguru==0.7.3
Mako==1.3.11
MarkupSafe==3.0.3
packaging==26.1
passlib==1.7.4
pdf2image==1.17.0
pillow==12.2.0
proto-plus==1.27.2
protobuf==5.29.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
PyMuPDF==1.27.2.2
pyparsing==3.3.2
pytesseract==0.3.13
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.33.1
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.49
starlette==1.0.0
text-unidecode==1.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
win32_setctime==1.2.0
wrapt==2.1.2
